requirements.txt
absl-py==2.1.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
executing==2.2.0
fastdtw==0.3.4
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.0.3
Flask-Session==0.8.0
Flask-Sessions==0.1.5
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.7
imbalanced-learn==0.12.3
imblearn==0.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.5.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mido==1.3.3
mistune==3.1.1
ml-dtypes==0.3.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
music21==9.3.0
mysql-connector-python==9.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.0
pooch==1.8.2
pretty_midi==0.2.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==4.25.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.0.8
referencing==0.36.2
regex==2024.7.24
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.22.3
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spotipy==2.25.0
SQLAlchemy==2.0.32
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
WTForms==3.1.2
xgboost==2.1.1
xlrd==2.0.1
